# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/admin/doctors/{doctorId}/slots": {
      "post": {
        "operationId": "AdminAppointmentsController_createSlot",
        "parameters": [
          {
            "name": "doctorId",
            "required": true,
            "in": "path",
            "description": "Doctor id (cuid)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSlotDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Slot created."
          },
          "400": {
            "description": "Past-time slot, endsAt <= startsAt, or DEACTIVATED doctor/category."
          },
          "401": {
            "description": "No active session."
          },
          "403": {
            "description": "Requires the admin role."
          },
          "404": {
            "description": "Doctor not found."
          }
        },
        "security": [
          {
            "vezeta.session_token": []
          }
        ],
        "summary": "Create a slot for a doctor (admin)",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/admin/slots": {
      "get": {
        "operationId": "AdminAppointmentsController_listSlots",
        "parameters": [
          {
            "name": "doctorId",
            "required": false,
            "in": "query",
            "description": "Filter by doctor id.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Filter by status.",
            "schema": {
              "type": "string",
              "enum": [
                "AVAILABLE",
                "BOOKED",
                "BLOCKED"
              ]
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "1-based page number.",
            "schema": {
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "description": "Items per page (1-100).",
            "schema": {
              "default": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of slots."
          },
          "401": {
            "description": "No active session."
          },
          "403": {
            "description": "Requires the admin role."
          }
        },
        "security": [
          {
            "vezeta.session_token": []
          }
        ],
        "summary": "List slots (admin)",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/admin/slots/{id}": {
      "get": {
        "operationId": "AdminAppointmentsController_getSlot",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Slot id (cuid)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Slot found."
          },
          "401": {
            "description": "No active session."
          },
          "403": {
            "description": "Requires the admin role."
          },
          "404": {
            "description": "Slot not found."
          }
        },
        "security": [
          {
            "vezeta.session_token": []
          }
        ],
        "summary": "Get a slot by id (admin)",
        "tags": [
          "admin"
        ]
      },
      "patch": {
        "operationId": "AdminAppointmentsController_updateSlot",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Slot id (cuid)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSlotDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Slot updated."
          },
          "400": {
            "description": "No fields to update."
          },
          "401": {
            "description": "No active session."
          },
          "403": {
            "description": "Requires the admin role."
          },
          "404": {
            "description": "Slot not found."
          }
        },
        "security": [
          {
            "vezeta.session_token": []
          }
        ],
        "summary": "Update a slot (partial)",
        "tags": [
          "admin"
        ]
      },
      "delete": {
        "operationId": "AdminAppointmentsController_deleteSlot",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Slot id (cuid)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Slot deleted."
          },
          "401": {
            "description": "No active session."
          },
          "403": {
            "description": "Requires the admin role."
          },
          "404": {
            "description": "Slot not found."
          },
          "409": {
            "description": "Slot is BOOKED or BLOCKED — only AVAILABLE slots can be deleted."
          }
        },
        "security": [
          {
            "vezeta.session_token": []
          }
        ],
        "summary": "Hard-delete an AVAILABLE slot (admin)",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/admin/slots/{id}/block": {
      "patch": {
        "operationId": "AdminAppointmentsController_blockSlot",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Slot id (cuid)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Slot blocked."
          },
          "401": {
            "description": "No active session."
          },
          "403": {
            "description": "Requires the admin role."
          },
          "404": {
            "description": "Slot not found."
          },
          "409": {
            "description": "Slot is BOOKED."
          }
        },
        "security": [
          {
            "vezeta.session_token": []
          }
        ],
        "summary": "Soft-block a slot (idempotent)",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/admin/appointments/{id}/confirm": {
      "patch": {
        "operationId": "AdminAppointmentsController_confirmAppointment",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Appointment id (cuid)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Appointment confirmed."
          },
          "401": {
            "description": "No active session."
          },
          "403": {
            "description": "Requires the admin role."
          },
          "404": {
            "description": "Appointment not found."
          },
          "409": {
            "description": "Appointment is not in PENDING status."
          }
        },
        "security": [
          {
            "vezeta.session_token": []
          }
        ],
        "summary": "Confirm a PENDING appointment (admin)",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/admin/appointments/{id}/cancel": {
      "patch": {
        "operationId": "AdminAppointmentsController_cancelAppointment",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Appointment id (cuid)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Appointment cancelled; slot released."
          },
          "401": {
            "description": "No active session."
          },
          "403": {
            "description": "Requires the admin role."
          },
          "404": {
            "description": "Appointment not found."
          },
          "409": {
            "description": "Appointment is already CANCELLED or COMPLETED."
          }
        },
        "security": [
          {
            "vezeta.session_token": []
          }
        ],
        "summary": "Cancel any appointment (admin)",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/admin/appointments/{id}/complete": {
      "patch": {
        "operationId": "AdminAppointmentsController_completeAppointment",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Appointment id (cuid)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Appointment completed."
          },
          "400": {
            "description": "scheduledAt is in the future."
          },
          "401": {
            "description": "No active session."
          },
          "403": {
            "description": "Requires the admin role."
          },
          "404": {
            "description": "Appointment not found."
          },
          "409": {
            "description": "Appointment is not in CONFIRMED status."
          }
        },
        "security": [
          {
            "vezeta.session_token": []
          }
        ],
        "summary": "Mark a CONFIRMED appointment as completed (admin)",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/admin/appointments": {
      "get": {
        "description": "Paginated list of all appointments, filterable by status, userId, doctorId.",
        "operationId": "AdminAppointmentsController_listAppointments",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Filter by status.",
            "schema": {
              "type": "string",
              "enum": [
                "PENDING",
                "CONFIRMED",
                "COMPLETED",
                "CANCELLED"
              ]
            }
          },
          {
            "name": "userId",
            "required": false,
            "in": "query",
            "description": "Filter by patient user id.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "doctorId",
            "required": false,
            "in": "query",
            "description": "Filter by doctor id.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "1-based page number.",
            "schema": {
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "description": "Items per page (1-100).",
            "schema": {
              "default": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of appointments."
          },
          "401": {
            "description": "No active session."
          },
          "403": {
            "description": "Requires the admin role."
          }
        },
        "security": [
          {
            "vezeta.session_token": []
          }
        ],
        "summary": "List all appointments (admin)",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/admin/appointments/{id}": {
      "get": {
        "operationId": "AdminAppointmentsController_getAppointment",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Appointment id (cuid)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Appointment found."
          },
          "401": {
            "description": "No active session."
          },
          "403": {
            "description": "Requires the admin role."
          },
          "404": {
            "description": "Appointment not found."
          }
        },
        "security": [
          {
            "vezeta.session_token": []
          }
        ],
        "summary": "Get an appointment by id (admin)",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/doctors/{doctorId}/slots": {
      "get": {
        "description": "Anonymous-accessible list of AVAILABLE slots for an ACTIVE doctor in an ACTIVE category, sorted ascending by start time.",
        "operationId": "SlotsController_listPublicSlots",
        "parameters": [
          {
            "name": "doctorId",
            "required": true,
            "in": "path",
            "description": "Doctor id (cuid)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Available slots, sorted ascending.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "slots": {
                      "type": "array",
                      "items": {
                        "type": "object"
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Doctor not found, is DEACTIVATED, or is in a DEACTIVATED category."
          },
          "429": {
            "description": "Rate limit exceeded."
          }
        },
        "summary": "List AVAILABLE slots for a doctor (public)",
        "tags": [
          "slots"
        ]
      }
    },
    "/api/appointments": {
      "post": {
        "operationId": "AppointmentsController_book",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookAppointmentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Appointment created in PENDING. Slot is now BOOKED."
          },
          "400": {
            "description": "Past-time slot, DEACTIVATED doctor or category."
          },
          "401": {
            "description": "No active session."
          },
          "403": {
            "description": "Account is deactivated."
          },
          "404": {
            "description": "Slot does not exist."
          },
          "409": {
            "description": "Slot is no longer available."
          },
          "429": {
            "description": "Rate limit exceeded."
          }
        },
        "summary": "Book a slot (patient)",
        "tags": [
          "appointments"
        ]
      },
      "get": {
        "operationId": "AppointmentsController_listMine",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Filter by status.",
            "schema": {
              "type": "string",
              "enum": [
                "PENDING",
                "CONFIRMED",
                "COMPLETED",
                "CANCELLED"
              ]
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "1-based page number.",
            "schema": {
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "description": "Items per page (1-100).",
            "schema": {
              "default": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of the patient's appointments."
          },
          "400": {
            "description": "Invalid query parameters."
          },
          "401": {
            "description": "No active session."
          }
        },
        "summary": "List my appointments (patient)",
        "tags": [
          "appointments"
        ]
      }
    },
    "/api/appointments/{id}/cancel": {
      "patch": {
        "operationId": "AppointmentsController_cancelMine",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Appointment cancelled; slot released."
          },
          "401": {
            "description": "No active session."
          },
          "403": {
            "description": "Within 24h of scheduled time."
          },
          "404": {
            "description": "Appointment does not exist or belongs to a different patient."
          },
          "409": {
            "description": "Appointment is already CANCELLED or COMPLETED."
          }
        },
        "summary": "Cancel my appointment (patient)",
        "tags": [
          "appointments"
        ]
      }
    },
...
  },
  "info": {
    "title": "Vezeeta Backend API",
    "description": "REST API for the Vezeeta backend.\n\n## Auth\n\nAuth is handled by [Better Auth](https://www.better-auth.com/) and exposed under `/api/auth/*`.\nAfter `sign-in/email` or `sign-in/phone-number`, the server sets an HTTP-only session cookie.\nThe frontend must send the cookie on every subsequent request (use `credentials: \"include\"` in fetch / `withCredentials: true` in axios).\n\nFor the social OAuth flow, the browser hits `GET /api/auth/oauth/start?provider=google` (or `facebook`) — a Better Auth standard route — and follows the 302 redirect.\n\nUse the **Authorize** button below to paste the session cookie value if you want to call protected endpoints from Swagger UI.\n\n## Tags\n\n- `auth` — current-session, social linking, health\n- `doctors` — public doctor catalog\n- `categories` — public category dropdown\n- `slots` — public slot picker for a doctor\n- `doctor-services` — per-doctor service catalog (admin)\n- `appointments` — patient booking flow\n- `reviews` — patient reviews + public doctor rating\n- `medical-records` — patient-facing read of their records\n- `notifications` — in-app inbox\n- `admin` — admin-only management (requires the `admin` role)",
    "version": "0.0.1",
    "contact": {},
    "x-generated-at": "2026-07-22T17:35:45.274Z"
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "vezeta.session_token",
        "description": "Better Auth session cookie set by /api/auth/* (prefix configured in src/auth/auth.ts)."
      }
    },
    "schemas": {
...
      "CreateSlotDto": {
        "type": "object",
        "properties": {
          "startsAt": {
            "format": "date-time",
            "type": "string",
            "description": "Slot start time (UTC, ISO 8601). Must be in the future.",
            "example": "2026-08-01T09:00:00.000Z"
          },
          "endsAt": {
            "format": "date-time",
            "type": "string",
            "description": "Slot end time (UTC, ISO 8601). Must be strictly after startsAt (service-layer check).",
            "example": "2026-08-01T09:30:00.000Z"
          }
        },
        "required": [
          "startsAt",
          "endsAt"
        ]
      },
      "UpdateSlotDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "description": "New slot status. BOOKED is NOT settable via this endpoint — booking manages that lifecycle.",
            "enum": [
              "AVAILABLE",
              "BLOCKED"
            ]
          }
        }
      },
      "BookAppointmentDto": {
        "type": "object",
        "properties": {
          "slotId": {
            "type": "string",
            "description": "Id of an AVAILABLE slot for an ACTIVE doctor in an ACTIVE category.",
            "maxLength": 64,
            "example": "slot_abc"
          },
          "patientNotes": {
            "type": "string",
            "description": "Patient-supplied context (e.g. reason for visit).",
            "maxLength": 2000
          }
        },
        "required": [
          "slotId"
        ]
      },
...
    }
  }
}